Solution found
This was a simple problem, and I decided to make it much more
complicated by adding in some concepts that I have yet to play with in
rust.

- Traits
- Implementing a trait for a existing/built-in type
- the `?` macro and returning a Result object
- using include_bytes! instead of reading a file

Traits were pretty straightforward. Nothing really interesting there
other than learning the syntax and how to declare the lifetimes to make
the code using them happy. I still have more to learn here, but I'm
getting stuff to compile and run.

The ? macro was something I wanted to use. I struggled with how to get
the types to line up when there were different types of errors in the
same function. While that's not the case now, since I'm using an
`unwrap_or` for the option, it was an issue when I was attempting to use
the macro on the option type. I played writing a impl for the From for
the various types, but that's not possible as the types are not defined
in this module. I came accross the map_err function on the result, which
would allow me to write a small block to convert the error. That seemed
to do what I want and would make this not a problem if I needed to do
this in the future.

I used include_bytes! for this file usage. Reading a file at runtime
made the code require different paths for when running in sublime's
build command and on the command line (for run and tests). Using this
made them all work the same, as the compiler includes the data into the
data section of the binary. It feels a little weird, but for something
like this where the file contents never change I don't really care about
the "purity" of it all.